src/app-server/backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
flask-cors
python-dotenv
gunicorn
requests